eris-app/agent-search/requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
langchain-google-genai>=2.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
psycopg2-binary>=2.9.9
pydantic>=2.5.0
python-dotenv>=1.0.0


